Made Yaks and fixed Layout bug
These will look a whole lot cleaner when I get the styling figured out,
but here's a prototype. Done with huge blocks of html for each Yak. Will
change to partial views for next iteration.

Also fixed a bug where the "PikYak" button in the navbar would redirect
to a nonexistent Home page. Now redirects back to the Feed.